Bump to 5.3.0

+++ labelme/widgets/canvas.py
@@ -561,7 +561,7 @@ class Canvas(QtWidgets.QWidget):
         y1 = top - point.y()
         x2 = right - point.x()
         y2 = bottom - point.y()
-        self.offsets = QtCore.QPoint(x1, y1), QtCore.QPoint(x2, y2)
+        self.offsets = QtCore.QPoint(int(x1), int(y1)), QtCore.QPoint(int(x2), int(y2))
 
     def boundedMoveVertex(self, pos):
         index, shape = self.hVertex, self.hShape
@@ -648,7 +648,7 @@ class Canvas(QtWidgets.QWidget):
         # Try to move in one direction, and if it fails in another.
         # Give up if both fail.
         point = shapes[0][0]
-        offset = QtCore.QPoint(2.0, 2.0)
+        offset = QtCore.QPoint(2, 2)
         self.offsets = QtCore.QPoint(), QtCore.QPoint()
         self.prevPoint = point
         if not self.boundedMoveShapes(shapes, point - offset):
@@ -724,7 +724,7 @@ class Canvas(QtWidgets.QWidget):
         aw, ah = area.width(), area.height()
         x = (aw - w) / (2 * s) if aw > w else 0
         y = (ah - h) / (2 * s) if ah > h else 0
-        return QtCore.QPoint(x, y)
+        return QtCore.QPoint(int(x), int(y))
 
     def outOfPixmap(self, p):
         w, h = self.pixmap.width(), self.pixmap.height()
@@ -769,11 +769,11 @@ class Canvas(QtWidgets.QWidget):
             # Handle cases where previous point is on one of the edges.
             if x3 == x4:
                 return QtCore.QPoint(
-                    min(max(0, x2), size.width() - 1),
-                    min(max(0, y2), max(y3, y4)))
+                    min(max(0, int(x2)), size.width() - 1),
+                    min(max(0, int(y2)), max(int(y3), int(y4))))
             else:  # y3 == y4
-                return QtCore.QPoint(min(max(0, x2), max(x3, x4)), y3)
-        return QtCore.QPoint(x, y)
+                return QtCore.QPoint(min(max(0, int(x2)), max(int(x3), int(x4))), int(y3))
+        return QtCore.QPoint(int(x), int(y))
 
     def intersectingEdges(self, point1, point2, points):
         """Find intersecting edges.
@@ -800,8 +800,8 @@ class Canvas(QtWidgets.QWidget):
             if 0 <= ua <= 1 and 0 <= ub <= 1:
                 x = x1 + ua * (x2 - x1)
                 y = y1 + ua * (y2 - y1)
-                m = QtCore.QPoint((x3 + x4) / 2, (y3 + y4) / 2)
-                d = labelme.utils.distance(m - QtCore.QPoint(x2, y2))
+                m = QtCore.QPoint(int((x3 + x4) / 2), int((y3 + y4) / 2))
+                d = labelme.utils.distance(m - QtCore.QPoint(int(x2), int(y2)))
                 yield d, i, (x, y)
 
     # These two, along with a call to adjustSize are required for the
